Dev Container Setup (#11)
Completed a full devcontainer setup. The `Dockerfile` should install all
dependencies except for the python dependencies. This is because I was
unable to figure out how to copy the requirements.txt file into the
DevContainer environment and run `pip install -r requirements.txt`. For
now the user will need run that command, which I think will need to be
done every time the container is rebuilt. We can consider putting this
as an issue to be fixed later.

This PR also contains the starter files for the Next.js app on the
frontend.

Some more workspace TODOs (possibly on another branch so we can start
development):

- [ ] Set up the Django API
- [ ] Set up the Caddy webserver with proxy to allow
`localhost:port/api/` routes.
- [ ] Configure scripting (possibly Honcho) to allow for a one-command
run
- [ ] Set up Prettier and a python formatter to ensure consistent code
styling

Also note the setup instructions found in `README.md`