Full Backend and Dev Environment Setup (#17) (#18)
Changed the stack to use React on the frontend and FastAPI with SQLModel
on the backend interacting with a SQLite database.

The "count" feature is the example feature that was setup. Please refer
to the `count` files found in each of the `api`, `services`, `models`,
and `testing/fake_data` folders.

The database is setup in the `reset_database` script found in the
`script` folder.

To work with the databse, you can either create your own session using
the `engine` found in the `db.py` file, or use FastAPI's dependency
injection (preferred) using the `db_session` also found in the `db.py`
file. Simply use `Depends(db_session)` as modeled in the
`services/count` file.

Additionally modified the DevContainer to not include a few extensions
and include a SQLite database explorer that you can use by clicking on
the `database.db` file that is generated when running the
`reset_database` script.

---------

Co-authored-by: Mustafa Aljumayli <idmustafase@gmail.com>