Refactor TeamService and Team Generation scripts
This is a big commit, but it's mostly refactoring. The team generation
scripts have been refactored so that it is easier to understand and
maintain. Originally, the TeamService had to interact with both
dataframes and Team objects, but now it only interacts with Team unless
specifically unpackaging or exporting a dataframe. This makes the code
more modular.

I also added a PasswordService to handle password generation so that
the TeamService doesn't have to worry about it.

WHAT STILL NEEDS TO BE DONE: Update Readme to reflect new scripts and
functionality.