DevContainer auto installs python dependencies now
All I needed to do was have the DevContainer run the Dockerfile script
in the parent directory, this was done via the
`{"build": {"context": ".."}}` (".." instead of ".") inside of the
devcontainer.json file.

The stuff in the Dockerfile simply copys the current requirements.txt
to the docker filesystem and executes off of it. Now we don't need to
run that annoying python3 -m pip install... blah blah blah yuck