The backend dependencies now automatically install on devcontainer build! (#23)
Adjusted a few files (see commit messages) in the docker setup to now
allow the backend dependencies to be installed when the devcontainer is
built. So as long as the `requirements.txt` is kept up to date via:

> anytime you install a new python package, "save" it to the
`requirements.txt` via `python3 -m pip freeze >
backend/requirements.txt`

Then anytime we pull changes and notice we're missing a package all we
need to do is rebuild the devcontainer! (`Ctrl/Cmd + P` and run `Dev
Containers: Rebuild and Reopen Container`) then everything will be setup
for us! (You can also just install that package lamo, but this is for
those working after us)