Accounts (#24)
Hello! Sorry, this is long overdue.

This PR has a few crucial updates to look into:

## Team Model and Service
This PR is all about team management. In the changes, a Team pydantic
model is added with specifications according to our database schema. See
models/team.py for the fields and types.

### The TeamService class
There are a ton of basic and more complicated methods that assist with
team management in this service right now. These allow for the creation
editing and retrieval of teams in the standard ways. Additionally, this
service allows for the conversion between polars data frames/rows and a
list of teams. The reason for this is the new `load_teams` script that
is a functional requirement for this product.

## Team Generation via load_teams.py
Currently, team generation and updating is done through this script. the
load_teams script serves as a temporary interface between the ES and the
team table in the database. If a team needs to be updated, the ES can do
it from the teams csv file and push it into the database!

## Password Generation via PasswordService
This service is just a way to separate password manipulation and
creation from the TeamService (it was getting crowded...)

## Updated ReadMe
Documentation for ES / admin side scripts is now found in the ReadMe.

## Next steps...
### GenerateTeams
This script will generate a dummy teams.csv file (Or add teams if the
file exists) for the ES to use for team management.

---------

Co-authored-by: Andrew Lockard <alockard@unc.edu>
Co-authored-by: Mustafa Aljumayli <idmustafase@gmail.com>
Co-authored-by: Nick UNC <nickunc@nicholass-mbp-5.wireless-1x.unc.edu>