Fixed CORS Issues
Hey Matthew! The only change needed to get around the CORS issues was to
not input the full URL, just the path to the API! Everything else that
we have set up takes care of that <3

I also removed the unique_words file as that's in the gitignore now.
I also added some weird auto generated file to the .gitignore for the
frontend.