Add Validation Helper Method and Update Members
Added `authed_team` to api/auth to allow for easily getting the team
from the JWT via a Depends in the route parameter.

Updated the TeamMembers routes to take advantage of this. Also created
sample data.

Finally, this commit also adds Public and Create models to the Team
table. The Public model is extremly important so we can prevent sending
the password as plaintext.