Start Timer First Implementation
Linked together the home page and login page, with automatic reroutes.
Home page will route to login page if there is not a valid token,
otherwise it will show a countdown until the start of the test.

In order to get the time working correctly, I had to change the time
objects in the backend to be datetime objects. This had minor
overhead; however, I have a feeling it breaks Nick's loading script
so I placed a TODO there.