Mustafa submission component 2 (#54)
I finally have my feature in somewhat of a pushable state, however, it
hasn't been completed just yet. Here is what I have gotten done so far:

- The question page calls on Nick's API to retrieve the questions from
the backend.
- The questions page also has clickable questions that have a state set
so it can be passed into my widget.
- The submission widget component has a submission and docs tab that
takes in props from the question component and update question specific
documentation.
- Added the tab to switch between global and question specific docs.
- I display the markdown for question-specific docs in a separate tab
and the global docs is linked to the Python documentation's index.html
file. Global_documents isn't retrieved from the backend and is displayed
on the frontend.

Here is what's next for me:

Add the code-editor IDE so that people can drop some code in here with
syntax highlighting. - Task Difficulty: Medium

Begin integrating Judge0 and try to build an API request to export code
to the backend and from the backend to Judge0 and all the way back to
the frontend. Effectively handling code submissions. - Task Difficulty:
Hard

---------

Co-authored-by: Matthew Futch <111521614+mattfutch01@users.noreply.github.com>