Running Submissions (#69)
Sorry for the delay in getting this PR up. I was in DC over break and
had a paper due this morning at 10 in class :( Mustafa and I worked on
these changes together.

The submission widget allows students to upload python files and
write/run code and view the output. The code is saved per question so
the user can leave a question and come back with their progress saved.
The output also switches to the last run submission so its not confusing
anymore. I also included the changes on Question.tsx so markdown on the
question page displays better with strikethroughs, tables, etc.